add model and supporting crystal images
The model provides the backend abstraction for the game.

Specifically, I've coded a Game object which knows about wins, losses, and scores.

I subclass this to create a CrystalCollector object that knows about game play specific to Crystal Collector.

I'm also implementing a Crystal object that associates an img asset with a value randomizer to associate with each crystal.  These are aggregated into an array of 4 crystals owned by the CrystalCollector object.