add game controller
The controller knows about the view and the model and interacts with both.

Most of the view will be static html, but the clickable crystals themselves are generated on the fly with jquery (to meet the intent of the assignment) and are associated with a click listener callback.

Beyond that, the controller reaches into the game model to get state which it then pushes to the view.

It also connects the view and model through a data-index attribute on the view element which is used to index into an array of crystal objects owned by the model to fetch the particular random value associated with each crystal so the current score can be updated appropriately.