avoid code duplication by calling superclass method from overriding subclass method
I was duplicating code in the CrystalCollector.reset() method.  But js, like other languages,
provides the option of invoking a superclass method with a couple syntactic options:

I'm opting for the more traditiona invocation since that seems to match the style of inheritance
I'm currently using:

CrystalCollector.prototype.reset = function() {
  Game.prototype.reset.call(this); // :-) call superclass reset() method
  for (let index in this.crystals) {
    this.crystals[index].reset();
  }
  this.targetValue = this.getRandomValue();
};

While I'm at it, I'll update the readme with this interesting tale since it amounted to a bug
I encountered during integration testing.